% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operators.R
\name{\%_\%}
\alias{\%_\%}
\title{Binary paste operator}
\usage{
a \%_\% b
}
\arguments{
\item{a}{An \code{R} object, to be converted to a character vector}

\item{b}{An \code{R} object, to be converted to a character vector}
}
\value{
A character vector of the concatenated values.
}
\description{
A binary version of the \code{\link[base]{paste0}} function.
}
\examples{
'I am ' \%_\% 31 \%_\% ' years old.'

}
