% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operators.R
\name{\%left\%}
\alias{\%left\%}
\alias{\%right\%}
\title{Binary operator to extract the first or last few characters of a string}
\usage{
vector \%left\% nr_chars

vector \%right\% nr_chars
}
\arguments{
\item{vector}{A character vector from which substrings will be extracted}

\item{nr_chars}{An integer number of characters to be extracted.}
}
\value{
A character vector containing substrings of \code{vector}.
}
\description{
Extracts the leftmost (\verb{\%left\%}) or rightmost (\verb{\%right\%}) few characters
of character strings.
}
\examples{
c('text', 'vec') \%left\% 4
c('text', 'vec') \%right\% 2

}
