% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operators.R
\name{\%like\%}
\alias{\%like\%}
\title{Binary pattern-matching operator}
\usage{
vector \%like\% pattern
}
\arguments{
\item{vector}{A character vector where matches are sought, or an object which
can be coerced by \code{as.character} to a character vector.}

\item{pattern}{A character string containing a
\link[base]{regular expression} to be matched in the given character vector.}
}
\value{
A logical vector (match or not for each element of \code{vector}).
}
\description{
A binary version of the \code{\link[base:grep]{base::grepl()}} function.
Determines if a character string matches a specified pattern.
Similar to SQL's \code{LIKE} operator, but uses \link[base]{regular expression}
syntax instead of SQL's wildcard characters * and \%.
}
\examples{
c('Hello world!', '12345') \%like\% '[a-z]'

}
