% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operators.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{\%!in\%}
\alias{\%not_in\%}
\title{Negated \code{\%in\%} operator}
\usage{
x \%notin\% vector

x \%!in\% vector

x \%not_in\% vector
}
\arguments{
\item{x}{Vector or \code{NULL}: the values to be matched}

\item{vector}{Vector or \code{NULL}: the values to be matched against}
}
\value{
A logical vector of the same length as \code{x}; returns \code{TRUE} when the
element of \code{x} is absent from \code{vector} and \code{FALSE} when the element of \code{x} is
present in \code{vector}.
}
\description{
Returns the logical negation of the binary \code{\link[base]{match}}.
Aliases: \verb{\%!in\%} and \verb{\%not_in}.
}
\examples{
c('text', 'vec') \%notin\% c('ice-cream', 'world', 'text')

}
