% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operators.R
\name{\%+\%}
\alias{\%+\%}
\title{Binary paste operator}
\usage{
a \%+\% b
}
\arguments{
\item{a}{An \code{R} object, to be converted to a character vector}

\item{b}{An \code{R} object, to be converted to a character vector}
}
\value{
A character vector of the concatenated values.
}
\description{
A binary version of the \code{\link[base]{paste0}} function.
}
\examples{
'I am ' \%+\% 31 \%+\% 'years old.'

}
