% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{progress_bar_start}
\alias{progress_bar_start}
\alias{progress_bar_tick}
\title{Make a simple progress bar in the terminal}
\usage{
progress_bar_start(repeats)

progress_bar_tick(beep = FALSE)
}
\arguments{
\item{repeats}{Integer. Number of ticks (events that move the progress bar
forward)}

\item{beep}{Logical. If \code{TRUE}, a sound will be played upon completion.}
}
\description{
Simple wrapper around \code{\link[progress:progress_bar]{progress::progress_bar()}} that sacrifices
the amount of options in favor of being easier to use. The progress bar
displays percent completion, time elapsed, and estimated time to conclusion.
Optionally, a sound may be played when the progress bar is finished (from
\code{\link[beepr:beep]{beepr::beep()}}).
}
\details{
Calling \code{progress_bar_start} initiates a progress bar. Calling
\code{progress_bar_tick} updates the progress. Should the argument \code{beep} be set
to \code{TRUE}, a sound will be played upon completion.
}
\examples{
progress_bar_start(repeats = 10)
for(i in 1:10) {
 progress_bar_tick()
 Sys.sleep(0.1);
}

}
