% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical. Should missing values be removed?}
}
\description{
Computes the standard error of the values in \code{x}. By default removes missing
values before the computation proceeds (this behavior is the opposite of
base R functions).
}
\details{
Like \code{\link[stats:cor]{stats::var()}} this uses denominator $n-1$.
The standard error of a length-one or zero-length vector is \code{NA}.
}
\examples{
se(c(1:5, NA))^2
se(c(1:5, NA), na.rm = FALSE)^2

}
