% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{tic}
\alias{tic}
\alias{toc}
\title{Record the time difference between the execution of two lines of code}
\usage{
tic(name = "timer")

toc(name = "timer")
}
\arguments{
\item{name}{Character string specifying the name of the timer. Optional, but
necessary for running multiple timers.}
}
\description{
Computes the time difference between a call to \code{tic} and a call to \code{toc}.
}
\details{
Calling \code{tic} initiates a timer. After initiating the timer, \code{toc}
outputs the time difference. \code{toc} can be called multiple times. The
timer can be reset by calling \code{tic} again.

Multiple timers (including nested timers) are supported: simply provide a
matching \code{name} argument to \code{tic} and to \code{toc}.
}
\examples{

# Example where `toc()` is called multiple times on the same timer:
tic()
print("Hello...")
Sys.sleep(1)
toc()
print("... world")
Sys.sleep(1)
toc()

# Example with nested timers:
tic('outer')
for(i in 1:3){
  tic('inner')
  Sys.sleep(0.2)
  toc('inner')
}
toc('outer')

}
